+++ src/main/java/com/ssafy/social/member/dto/request/AddGroupMemberRequest.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class AddGroupMemberRequest {
-    private int targetId;  // 초대할 사용자 id
+    private Long targetId;  // 초대할 사용자 id
 }
